requirements.txt
django==3.2.23
django-tagging==0.5.0
django-reversion==4.0.0
django-bootstrap3==21.2
django-summernote==0.8.11.6
django-modeltranslation==0.18.2
django-registration==3.1.2
django-contrib-comments==1.9.2
django-queryset-csv==1.1.0
django-debug-toolbar==3.2.2
django-guardian==2.2.0
django-notifications-hq==1.6.0
django-background-tasks==1.2.5
dj-database-url==0.5.0
boto3==1.22.8
django-storages==1.12.1
psycopg2==2.8.6
certifi==2023.7.22
urllib3==1.26.18
sentry-sdk==1.14.0
gunicorn==20.1.0
whitenoise==6.2.0
